% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_plot.R
\name{anno_plot}
\alias{anno_plot}
\title{Annotations ggplot2 plot}
\usage{
anno_plot(
  geom = "plot",
  x = NULL,
  y = NULL,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  xend = NULL,
  yend = NULL,
  na.rm = FALSE,
  label = NULL,
  ...
)
}
\arguments{
\item{geom}{character Name of geom to use for annotation.}

\item{x, y, xmin, ymin, xmax, ymax, xend, yend, }{numeric Positioning
aesthetics - you must specify at least one of these.}

\item{na.rm}{logical If \code{FALSE}, the default, missing values are removed
with a warning. If TRUE, missing values are silently removed.}

\item{label}{character, data.frame, ggplot or grob.}

\item{...}{Other named arguments passed on to \code{layer()}. These are often
aesthetics, used to set an aesthetic to a fixed value, like color = "red"
or size = 3. They may also be parameters to the paired geom/stat.}
}
\value{
A plot layer instance.
}
\description{
Annotations ggplot2 plot
}
\examples{
p <- ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point()
p + anno_plot("text", x = 5, y = 32, label = "Some text")
p + anno_plot("label", x = c(2, 5), y = c(15, 32), label = c("A", "B"))
p + anno_plot("plot", x = 5.5, y = 34, label = p + theme_bw(9))
p + anno_plot("rect", xmin = 3, xmax = 4.2, ymin = -Inf, ymax = Inf, alpha = .2, fill = "darkred")
p + anno_plot("segment", x = 2.5, xend = 4, y = 15, yend = 25, colour = "blue")
p + anno_plot("pointrange", x = 3.5, y = 20, ymin = 12, ymax = 28, colour = "red", size = 1.5)

}
