% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorplot.R
\name{colorplot}
\alias{colorplot}
\title{plot the colors with values}
\usage{
colorplot(type = "npg")
}
\arguments{
\item{type}{the type of journal}
}
\value{

}
\description{
plot the colors with values
}
\examples{
colorplot(type = "npg")
}
\author{
Zhonghui Gai
}
