% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.in.R
\name{data.in}
\alias{data.in}
\title{Read in the data for analysis by using the read.delim and read.xlsx functions}
\usage{
data.in(file.path = "./data.txt", colname = TRUE, rowname = TRUE, sheet = 1)
}
\arguments{
\item{file.path}{the file path to be read in}

\item{colname}{If TRUE, the first row of data will be used as column names.}

\item{rowname}{If TRUE, first column of data will be used as row names.}

\item{sheet}{The name or index of the sheet to read data from.}
}
\value{
a matrix or a data frame
}
\description{
Read in the data for analysis by using the read.delim and read.xlsx functions
}
\examples{
test <- data.in("sum_phylum.txt", rowname = FALSE)
test <- data.in("sum_phylum.txt")
test <- data.in("group.xlsx", rowname = F)
}
\author{
Zhonghui Gai
}
