% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.out.R
\name{data.out}
\alias{data.out}
\title{Save data using write.table function}
\usage{
data.out(file = "mtcars", colname = TRUE, rowname = FALSE)
}
\arguments{
\item{file}{the file to be saved, usually a data frame or a matrix}

\item{colname}{either a logical value indicating whether the column names of x are to be written along with x, or a character vector of column names to be written.}

\item{rowname}{either a logical value indicating whether the row names of x are to be written along with x, or a character vector of row names to be written.}
}
\value{

}
\description{
Save data using write.table function
}
\examples{
gfun::data.out(file = "mtcars")
}
\author{
Zhonghui Gai
}
