% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_long.R
\name{data_long}
\alias{data_long}
\title{Pivot data from wide to long}
\usage{
data_long(
  data,
  cols = "group",
  names_to = "name",
  names_prefix = NULL,
  names_sep = NULL,
  names_pattern = NULL,
  values_to = "value",
  values_drop_na = FALSE,
  ...
)
}
\arguments{
\item{data}{`data.frame`. The data to pivot.}

\item{cols}{<[`poor-select`][select_helpers]>. Columns to pivot into longer format.}

\item{values_to}{`character(n)`. The name of the new column(s) that will contain the values of the pivoted variables.}

\item{...}{Additional arguments passed on to methods.}
}
\value{
A `data.frame`.
}
\description{
`data_long()` "lengthens" data, increasing the number of rows and decreasing the number of columns. The inverse
transformation is [pivot_wider()].
}
\examples{
wide_data <- data.frame(replicate(5, rnorm(10)))
# Customizing the names
data_long(
  data = wide_data,
  cols = c(1, 2),
  names_to = "Column",
  values_to = "Numbers"
)

}
