% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_write.R
\name{data_write}
\alias{data_write}
\title{Save data using write.table function}
\usage{
data_write(data, rowname = FALSE, colname = TRUE)
}
\arguments{
\item{rowname}{either a logical value indicating whether the row names of x are to be written along with x, or a character vector of row names to be written.}

\item{colname}{either a logical value indicating whether the column names of x are to be written along with x, or a character vector of column names to be written.}

\item{file}{the file to be saved, usually a data frame or a matrix}
}
\description{
Save data using write.table function
}
\examples{
data_write(mtcars, T)
}
\author{
Zhonghui Gai
}
