% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.rm.R
\name{env.rm}
\alias{env.rm}
\title{Delete objects in the envirement using grepl function}
\usage{
env.rm(del = c("^Day", "Day9$"))
}
\arguments{
\item{del}{a vector containing the names of objects to be deleted}
}
\value{
a list
}
\description{
Delete objects in the envirement using grepl function
}
\examples{
Day6 <- 1
l <- ls()
rm(list = env.rm(c("^Day")))
rm(l)
}
\author{
Zhonghui Gai
}
