% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.rm.R
\name{env_hold}
\alias{env_hold}
\title{Hold objects in the envirement using grepl function}
\usage{
env_hold(hold = c("temp"))
}
\arguments{
\item{hold}{a vector containing the names of objects to be hold}
}
\description{
Hold objects in the envirement using grepl function
}
\examples{
Day6 <- 1
a_Day9 <- 2
temp <- 3
env_hold(hold = c("temp"))
}
\author{
ZhonghuiGai
}
