% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.rm.R
\name{env_rm}
\alias{env_rm}
\title{Delete objects in the envirement using grepl function}
\usage{
env_rm(del = c("^Day", "Day9$"))
}
\arguments{
\item{del}{a vector containing the names of objects to be deleted}
}
\description{
Delete objects in the envirement using grepl function
}
\examples{
Day6 <- 1
a_Day9 <- 2
temp <- 3
env_rm(del = c("^Day"))
}
\author{
Zhonghui Gai
}
