% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfunlist.R
\name{gfunlist}
\alias{gfunlist}
\title{Show the list of packages avilable}
\usage{
gfunlist()
}
\value{

}
\description{
Show the list of packages avilable
}
\examples{
gfun::gfunlist()
}
\author{
Zhonghui Gai
}
