% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglab.R
\name{gglab_Greek_bold_superscript}
\alias{gglab_Greek_bold_superscript}
\title{Greek, bold, and superscript symbols in ggplot2 labs}
\usage{
gglab_Greek_bold_superscript()
}
\value{

}
\description{
Greek, bold, and superscript symbols in ggplot2 labs
}
\examples{
gglab_Greek_bold_superscript()
}
\author{
Zhonghui Gai
}
