% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglab.R
\name{ggpoint_Greek_Cross}
\alias{ggpoint_Greek_Cross}
\title{Greek and Cross point shapes for ggplot2}
\usage{
ggpoint_Greek_Cross()
}
\value{

}
\description{
Greek and Cross point shapes for ggplot2
}
\examples{
ggpoint_Greek_Cross()
}
\author{
Zhonghui Gai
}
