% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_github.R
\name{install_github}
\alias{install_github}
\title{Install packages form github}
\usage{
install_github(pkg = "ggroup")
}
\arguments{
\item{pkg}{the package name}
}
\value{

}
\description{
Install packages form github
}
\examples{
gfun::install_github(:ggroup)
}
\author{
Zhonghui Gai
}
