% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheatmap.R
\name{pheatmap}
\alias{pheatmap}
\title{A swap function of pheatmap, the inputs can be one or two matix}
\usage{
pheatmap(
  mat,
  cluster_cols = T,
  cluster_rows = T,
  display_numbers = FALSE,
  color = c("#3c9eff", "gray99", "#ff445d"),
  treeheight_row = 20,
  treeheight_col = 20,
  cellwidth = NA,
  cellheight = NA,
  cutree_rows = 2,
  cutree_cols = 2,
  fontsize = 6,
  fontsize_number = 9,
  fontsize_row = 11,
  fontsize_col = 11,
  angle_col = 45,
  ...
)
}
\arguments{
\item{mat}{numeric matrix of the values to be plotted.}

\item{cluster_cols}{boolean values determining if columns should be clustered or hclust object.}

\item{cluster_rows}{boolean values determining if rows should be clustered or hclust object,}

\item{display_numbers}{logical determining if the numeric values are also printed to the cells. If this is a matrix (with same dimensions as original matrix), the contents of the matrix are shown instead of original values.}

\item{color}{vector of colors used in heatmap.}

\item{treeheight_row}{treeheight_row}

\item{treeheight_col}{treeheight_col}

\item{cellwidth}{cellwidth}

\item{cellheight}{cellheight}

\item{cutree_rows}{cutree_rows}

\item{cutree_cols}{cutree_cols}

\item{fontsize}{fontsize}

\item{fontsize_number}{fontsize_number}

\item{fontsize_row}{fontsize_row}

\item{fontsize_col}{fontsize_col}

\item{angle_col}{angle_col}

\item{...}{other param}
}
\description{
A swap function of pheatmap, the inputs can be one or two matix
}
\author{
Zhonghui Gai
}
