% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot as PDF}
\usage{
save_plot(width = 10, height = 10, plot = last_plot(), ...)
}
\arguments{
\item{width}{Width of the image in inches}

\item{height}{Height of the image in inches}

\item{plot}{A plot object}

\item{...}{other parameters}

\item{filename}{Filename to save to an PDF}
}
\description{
Save a plot as PDF
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
}
\author{
Zhonghui Gai
}
