% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_manual2.R
\name{scale_color_manual2}
\alias{scale_color_manual2}
\title{A wrapper function of set colors for ggplot2 manually}
\usage{
scale_color_manual2(
  type = "npg",
  order = c(3, 5, 6),
  values = NULL,
  breaks = waiver(),
  ...
)
}
\arguments{
\item{type}{the type of journal, one of npg aaas nejm lancet jama jco uchicago d3 ijv locus simpsons futurma richandorty startrek tron gsea pub}

\item{order}{the order and the number colors chooses for using}

\item{values}{a color vector to be used, when this values used the type will be ognined}

\item{breaks}{breaks}

\item{...}{other parameters to be set}
}
\value{

}
\description{
A wrapper function of set colors for ggplot2 manually
}
\examples{
ggplot(data = data.frame(x = as.factor(1:5), y = 1:5), aes(x,y)) + geom_point(aes(color = x)) +
gfun::scale_color_manual2(order = c(2,5,7,1,4))
library(ggplot2)
library(ggroup)
data <- iris
colnames(data)[5] <- "group"
colorplot(type = "npg")
ggbarplot(data = data, variable = "Petal.Width",
          comparisons = list(c("setosa", "versicolor"),
                             c("versicolor", "virginica"),
                             c("setosa", "virginica"))) +  gfun::scale_color_manual2(type = "npg", order = c(3,5,6))
}
\author{
Zhonghui Gai
}
