% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeplot.R
\name{shapeplot}
\alias{shapeplot}
\title{Show all the shape of points}
\usage{
shapeplot()
}
\value{
a ggplot2 obj
}
\description{
Show all the shape of points
}
\examples{
shapeplot()
}
\author{
Zhonghui Gai
}
