% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.color.R
\name{show.color}
\alias{show.color}
\title{A quick and dirty way to show colours in a base plot}
\usage{
show.color(colors, ncol = NULL)
}
\arguments{
\item{colors}{A character vector of colours}

\item{ncol}{Number of columns. If not supplied, tries to be as square as possible.}
}
\value{

}
\description{
A quick and dirty way to show colours in a base plot
}
\examples{
show.color(c("red", "green"))
}
\author{
adapt from scales package show_col function
}
