% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_vertical.R
\name{slice_vertical}
\alias{slice_vertical}
\title{Find columns from a data frame by pattern}
\usage{
slice_vertical(data, select, reverse = FALSE, drop = FALSE)
}
\arguments{
\item{data}{a date frame}

\item{select}{a list or a vector containing the select information, characters or numbers!}

\item{reverse}{a logical value to indicate whether to reverse the select, default value is FALSE}

\item{drop}{a logical value used in select only one column}
}
\value{
a data frame or a vector
}
\description{
Find columns from a data frame by pattern
}
\examples{
slice_vertical(data = iris, select = list("^Sepal"))
slice_vertical(data = iris, select = c("^Sepal", "Width$"))
slice_vertical(data = iris, select = c(3,5), reverse = TRUE)
slice_vertical(data = iris, select = c(3), reverse = FALSE, drop = TRUE)
slice_vertical(data = iris, select = c("\\\\."))
slice_vertical(data = mtcars, select = c("g"))  |> head()
}
\author{
Zhonghui Gai
}
