% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_pub}
\alias{theme_pub}
\title{ggplot theme for publication ready Plots}
\usage{
theme_pub(
  base_size = 14,
  border = FALSE,
  angle = 0,
  plot.margin = c(3, 3, 3, 3),
  panel.grid.major.color = "#f0f0f0"
)
}
\arguments{
\item{base_size}{the default value is 14}

\item{border}{`logical`. Should a border be drawn around the plot?}

\item{angle}{`integer`. Angle of x_axis text in degrees.
One of: `-30, 45, 90, 270`.}

\item{plot.margin}{plot.margin, trouble, mm}

\item{panel.grid.major.color}{the color of grid, default value is "#f0f0f0".}
}
\value{
ggtheme object
}
\description{
ggplot theme for publication ready Plots
}
\examples{
gglm::gglm() + ggtheme::theme_pub()
gglm(group = "Species", label.y = 0.99) + ggtheme::theme_pub() +
ggprism::scale_color_prism(palette = "floral")
}
\author{
Zhonghui Gai
}
