% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice}
\alias{slice}
\title{slice lets you index rows by a grouping variable, it's a wrapper function of grepl}
\usage{
slice(select = c("setosa$", "^virginica"), group = iris$Species, data = NULL)
}
\arguments{
\item{select}{a vector to be used for slice, character string containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector.}

\item{group}{the group variable, the column containing the grouping information}

\item{data}{a data frame or NULL}
}
\value{
a vector or data frame
}
\description{
slice lets you index rows by a grouping variable, it's a wrapper function of grepl
}
\examples{
ind <- slice(select = c("setosa$", "^virginica"), group = iris$Species, data = NULL)
iris.ind <- slice(select = c("setosa$", "^virginica"), group = iris$Species, data = iris)
data <- iris
colnames(data)[5] <- "group"
iris.ind2 <- slice(select = c("setosa$", "^virginica"), data = iris)
identical(iris.ind, iris.ind2)
subset(iris, Species == "setosa" | Species == "virginca")
slice(select = c(4,1), group = mtcars$carb, mtcars)
slice(select = c(21, 15), group = mtcars$mpg, mtcars)
}
\author{
Zhonghui Gai
}
